log file existence check

+++ project/mainqt.py
@@ -182,6 +182,9 @@ def main():
     application = ApplicationWindow()
 
     if len(sys.argv) > 2:
+        if os.path.isfile(sys.argv):
+            raise Exception("Log file of that name already exists")
+
         logfile = open(sys.argv[2], 'w')
     else:
         logfile = None
